read detection efficiencies from a normalization histogram and import them to PETSIRD header
Read and import detection efficiencies froma separate histogram file comprised of a list of entries with fllowing format:
detID1<uint32>, detID2<uint32>, detEff<float>
The detIDs were global and assigned first for all detectors of the first detector ring, then alll detectors of the second ring etc.
Currently assume only rotational symmetries when determining the module_pair_SGID_LUT. All detection efficiencies of detector pairs or LORs corresponding to the same symmetry group (SGID) were averaged. 
Functions created to calculate all layer IDs from the global detector index. Then the module indices and the detector element indices within the "PETSIRD module" (.e. the rsector in this case) were calculated. 

Note: We consider as the "PETSIRD module" the rsector layer here as an example. PETSIRD standard expects the user to choose which of the original layers of the ROOT will be considered as the "PETSIRD module". All layers below the "PETSIRD module" (i.e. the rsector in this case) are concatenated to a single layer representing the detector elements within the "PETSIRD module".